requirements.txt
flask==3.0.2
flask-session==0.5.0
pandas==2.2.0
pillow==10.2.0
requests==2.31.0
werkzeug==3.0.1
python-dotenv==1.0.1
openpyxl==3.1.2
xlrd==2.0.1
odfpy==1.4.1
matplotlib==3.8.2
numpy==1.26.3
scikit-learn==1.4.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
nltk==3.8.1 